moved locale patching methods to LanguageUtils class
